# src/main/java/br/com/bagnascojhoel/kwik_ecommerce/common/infrastructure/driving_ports/rest/resource_representations/FailedValidationRepresentation.java
package br.com.bagnascojhoel.kwik_ecommerce.common.infrastructure.driving_ports.rest.resource_representations;

import lombok.Builder;
import lombok.Getter;
import lombok.RequiredArgsConstructor;

@RequiredArgsConstructor
@Getter
@Builder
public class FailedValidationRepresentation {

    private final String field;
    private final String reason;

}